edit cabin

+++ src/services/apiCabins.js
@@ -2,6 +2,7 @@ import supabase, { supabaseUrl } from "./supabase";
 
 export async function getCabins() {
   const { data, error } = await supabase.from("cabins").select("*");
+
   if (error) {
     console.error(error);
     throw new Error("Cabins could not be loaded");
@@ -11,19 +12,24 @@ export async function getCabins() {
 }
 
 export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     ""
   );
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  //1. Create/edit cabin
-
+  // 1. Create/edit cabin
   let query = supabase.from("cabins");
-  if (!id) query.insert([{ ...newCabin, image: imagePath }]);
 
-  // b) EDIT
-  if (id) query.update({ ...newCabin, image: imagePath }).eq("id", id);
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
 
   const { data, error } = await query.select().single();
 
@@ -32,17 +38,19 @@ export async function createEditCabin(newCabin, id) {
     throw new Error("Cabin could not be created");
   }
 
-  //2. Upload image
+  // 2. Upload image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
 
-  //3. Delete the cabin if there was an error uploading image
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
     console.error(storageError);
     throw new Error(
-      "Cabins image could not uploaded and the cabin was not created"
+      "Cabin image could not be uploaded and the cabin was not created"
     );
   }
 
@@ -56,4 +64,6 @@ export async function deleteCabin(id) {
     console.error(error);
     throw new Error("Cabin could not be deleted");
   }
+
+  return data;
 }
